requirements.txt
cassandra-driver==3.29.0
redis==5.0.3
fastapi==0.104.1
uvicorn==0.23.2
